surfaceflinger: clean up captureScreen
dispScissor is never empty.  Remove the check and update the
comments so that the intention is clear.

It also becomes obvious that reqWidth/reqHeight is never zero when
we reach "// get screen geometry".  All code following the comment
is actually dead, which means we probably broke it since commit
06a58e22.

Bug: 113041375
Test: take screenshot, rotate screen, screencap
Change-Id: Ida1430383ce62271365d9ef64ad9b055638e1eac
Merged-In: Ida1430383ce62271365d9ef64ad9b055638e1eac
Signed-off-by: Josh Fox (XlxFoXxlX) <joshfox87@gmail.com>